requirements.txt
# Core
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.0
pydantic-settings>=2.0

# ML / Numerical
numpy>=1.26.0

# Inference backends
httpx>=0.27.0           # async HTTP client for Ollama / vLLM / OpenAI API
openai>=1.30.0          # OpenAI-compatible API (vLLM also exposes this)

# gRPC (edge <-> cloud communication)
grpcio>=1.60.0
grpcio-tools>=1.60.0
protobuf>=4.25.0

# Monitoring
prometheus-client>=0.20.0

# Testing
pytest>=8.0
pytest-asyncio>=0.23.0

# Dashboard
streamlit>=1.35.0
pandas>=2.0
plotly>=5.18.0

# Utilities
rich>=13.0              # pretty logging
